refactor: clean dont-consume-firework
