+++ src/client/java/com/koralix/oneforall/duck/DontConsumeFirework.java
@@ -1,0 +1,5 @@
+package com.koralix.oneforall.duck;
+
+public interface DontConsumeFirework {
+    void oneforall$DontConsumeFirework();
+}
