--- src/client/java/com/koralix/oneforall/duck/FireworkGetter.java
@@ -1,7 +1,0 @@
-package com.koralix.oneforall.duck;
-
-import net.minecraft.entity.projectile.FireworkRocketEntity;
-
-public interface FireworkGetter {
-    boolean isActiveFirework(FireworkRocketEntity firework);
-}
